requirements.txt
numpy
scipy 
tqdm 
PyOpenGL
plyfile
matplotlib
python-decouple
PyGLM
PyQt5